requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
boto3==1.9.201
botocore==1.12.201
chardet==3.0.4
docutils==0.14
idna==2.8
Jinja2==2.10.1
jmespath==0.9.4
MarkupSafe==1.1.1
multidict==4.5.2
python-dateutil==2.8.0
s3transfer==0.2.1
six==1.12.0
slackclient==2.1.0
urllib3==1.25.3
yarl==1.3.0
